Stubs For API

+++ etl-aro/etl/financial/create_financials.sql
@@ -376,8 +376,8 @@ create table financial.equipment_summary_cost (
 	
 	id bigserial primary key,
 	
-	network_cost_code_id int8 references financial.network_cost_code,
-	network_report_id int4 not null references financial.network_report on delete cascade ,
+	network_cost_code_id int4 references financial.network_cost_code,
+	network_report_id int8 not null references financial.network_report on delete cascade ,
 		
 	atomic_count double precision,
 	quantity double precision,
@@ -395,7 +395,7 @@ create table financial.fiber_item_cost (
 	
 	id bigserial primary key,
 	
-	network_cost_code_id int8 references financial.network_cost_code,
+	network_cost_code_id int4 references financial.network_cost_code,
 	network_report_id int8 not null references financial.network_report on delete cascade,
 	fiber_route_id int8 not null references client.fiber_route on delete cascade,
